Grading
You just missed the point of the prompt. There's essentially no difference between what you did in loading the data and having the dataset saved in your file. There is no part of your app which represents the concepts we went over in class.

There are also some funky parts of your UI, the chart placement could certainly be better.

Good luck in the rest of your semester!